% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb_budget}
\alias{tb_budget}
\title{Word Health Organization (WHO) Tuberculosis budgets by country}
\format{
A data frame with columns:
\describe{
\item{country}{The variable is character. }
\item{iso2}{The variable is character. }
\item{iso3}{The variable is character. }
\item{iso_numeric}{The variable is character. }
\item{g_whoregion}{The variable is character. }
\item{year}{The variable is numeric. }
\item{tx_dstb}{The variable is numeric. }
\item{budget_cpp_dstb}{The variable is numeric. }
\item{tx_mdr}{The variable is numeric. }
\item{budget_cpp_mdr}{The variable is numeric. }
\item{tx_xdr}{The variable is numeric. }
\item{budget_cpp_xdr}{The variable is numeric. }
\item{budget_lab}{The variable is numeric. }
\item{cf_lab}{The variable is numeric. }
\item{budget_staff}{The variable is numeric. }
\item{cf_staff}{The variable is numeric. }
\item{budget_fld}{The variable is numeric. }
\item{cf_fld}{The variable is numeric. }
\item{budget_prog}{The variable is numeric. }
\item{cf_prog}{The variable is numeric. }
\item{budget_sld}{The variable is numeric. }
\item{cf_sld}{The variable is numeric. }
\item{budget_mdrmgt}{The variable is numeric. }
\item{cf_mdrmgt}{The variable is numeric. }
\item{budget_tbhiv}{The variable is numeric. }
\item{cf_tbhiv}{The variable is numeric. }
\item{budget_patsup}{The variable is numeric. }
\item{cf_patsup}{The variable is numeric. }
\item{budget_orsrvy}{The variable is numeric. }
\item{cf_orsrvy}{The variable is numeric. }
\item{budget_oth}{The variable is numeric. }
\item{cf_oth}{The variable is numeric. }
\item{budget_tot}{The variable is numeric. }
\item{cf_tot}{The variable is numeric. }
\item{cf_tot_domestic}{The variable is numeric. }
\item{cf_tot_gf}{The variable is numeric. }
\item{cf_tot_usaid}{The variable is numeric. }
\item{cf_tot_grnt}{The variable is numeric. }
\item{cf_tot_sources}{The variable is numeric. }
}
}
\source{
\url{https://www.who.int/tb/country/data/download/en/}
}
\usage{
tb_budget
}
\description{
See source for description of the data. tb_dictionary describes the column names.
}
\examples{
\dontrun{
tb_budget
}
}
\keyword{datasets}
