% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb_cases}
\alias{tb_cases}
\title{Word Health Organization (WHO) Tuberculosis case notifications by country}
\format{
A data frame with columns:
\describe{
\item{country}{The variable is character. }
\item{iso2}{The variable is character. }
\item{iso3}{The variable is character. }
\item{iso_numeric}{The variable is character. }
\item{g_whoregion}{The variable is character. }
\item{year}{The variable is numeric. }
\item{new_sp}{The variable is numeric. }
\item{new_sn}{The variable is numeric. }
\item{new_su}{The variable is numeric. }
\item{new_ep}{The variable is numeric. }
\item{new_oth}{The variable is numeric. }
\item{ret_rel}{The variable is numeric. }
\item{ret_taf}{The variable is numeric. }
\item{ret_tad}{The variable is numeric. }
\item{ret_oth}{The variable is numeric. }
\item{newret_oth}{The variable is numeric. }
\item{new_labconf}{The variable is numeric. }
\item{new_clindx}{The variable is numeric. }
\item{ret_rel_labconf}{The variable is numeric. }
\item{ret_rel_clindx}{The variable is numeric. }
\item{ret_rel_ep}{The variable is numeric. }
\item{ret_nrel}{The variable is numeric. }
\item{notif_foreign}{The variable is numeric. }
\item{c_newinc}{The variable is numeric. }
\item{new_sp_m04}{The variable is numeric. }
\item{new_sp_m514}{The variable is numeric. }
\item{new_sp_m014}{The variable is numeric. }
\item{new_sp_m1524}{The variable is numeric. }
\item{new_sp_m2534}{The variable is numeric. }
\item{new_sp_m3544}{The variable is numeric. }
\item{new_sp_m4554}{The variable is numeric. }
\item{new_sp_m5564}{The variable is numeric. }
\item{new_sp_m65}{The variable is numeric. }
\item{new_sp_mu}{The variable is numeric. }
\item{new_sp_f04}{The variable is numeric. }
\item{new_sp_f514}{The variable is numeric. }
\item{new_sp_f014}{The variable is numeric. }
\item{new_sp_f1524}{The variable is numeric. }
\item{new_sp_f2534}{The variable is numeric. }
\item{new_sp_f3544}{The variable is numeric. }
\item{new_sp_f4554}{The variable is numeric. }
\item{new_sp_f5564}{The variable is numeric. }
\item{new_sp_f65}{The variable is numeric. }
\item{new_sp_fu}{The variable is numeric. }
\item{new_sn_m04}{The variable is numeric. }
\item{new_sn_m514}{The variable is numeric. }
\item{new_sn_m014}{The variable is numeric. }
\item{new_sn_m1524}{The variable is numeric. }
\item{new_sn_m2534}{The variable is numeric. }
\item{new_sn_m3544}{The variable is numeric. }
\item{new_sn_m4554}{The variable is numeric. }
\item{new_sn_m5564}{The variable is numeric. }
\item{new_sn_m65}{The variable is numeric. }
\item{new_sn_m15plus}{The variable is numeric. }
\item{new_sn_mu}{The variable is numeric. }
\item{new_sn_f04}{The variable is numeric. }
\item{new_sn_f514}{The variable is numeric. }
\item{new_sn_f014}{The variable is numeric. }
\item{new_sn_f1524}{The variable is numeric. }
\item{new_sn_f2534}{The variable is numeric. }
\item{new_sn_f3544}{The variable is numeric. }
\item{new_sn_f4554}{The variable is numeric. }
\item{new_sn_f5564}{The variable is numeric. }
\item{new_sn_f65}{The variable is numeric. }
\item{new_sn_f15plus}{The variable is numeric. }
\item{new_sn_fu}{The variable is numeric. }
\item{new_sn_sexunk04}{The variable is logical. }
\item{new_sn_sexunk514}{The variable is logical. }
\item{new_sn_sexunk014}{The variable is logical. }
\item{new_sn_sexunk15plus}{The variable is logical. }
\item{new_ep_m04}{The variable is logical. }
\item{new_ep_m514}{The variable is numeric. }
\item{new_ep_m014}{The variable is numeric. }
\item{new_ep_m1524}{The variable is numeric. }
\item{new_ep_m2534}{The variable is numeric. }
\item{new_ep_m3544}{The variable is numeric. }
\item{new_ep_m4554}{The variable is numeric. }
\item{new_ep_m5564}{The variable is numeric. }
\item{new_ep_m65}{The variable is numeric. }
\item{new_ep_m15plus}{The variable is numeric. }
\item{new_ep_mu}{The variable is numeric. }
\item{new_ep_f04}{The variable is numeric. }
\item{new_ep_f514}{The variable is numeric. }
\item{new_ep_f014}{The variable is numeric. }
\item{new_ep_f1524}{The variable is numeric. }
\item{new_ep_f2534}{The variable is numeric. }
\item{new_ep_f3544}{The variable is numeric. }
\item{new_ep_f4554}{The variable is numeric. }
\item{new_ep_f5564}{The variable is numeric. }
\item{new_ep_f65}{The variable is numeric. }
\item{new_ep_f15plus}{The variable is numeric. }
\item{new_ep_fu}{The variable is numeric. }
\item{new_ep_sexunk04}{The variable is numeric. }
\item{new_ep_sexunk514}{The variable is numeric. }
\item{new_ep_sexunk014}{The variable is numeric. }
\item{new_ep_sexunk15plus}{The variable is numeric. }
\item{new_ep_sexunkageunk}{The variable is logical. }
\item{rel_in_agesex_flg}{The variable is numeric. }
\item{newrel_m04}{The variable is numeric. }
\item{newrel_m514}{The variable is numeric. }
\item{newrel_m014}{The variable is numeric. }
\item{newrel_m1524}{The variable is numeric. }
\item{newrel_m2534}{The variable is numeric. }
\item{newrel_m3544}{The variable is numeric. }
\item{newrel_m4554}{The variable is numeric. }
\item{newrel_m5564}{The variable is numeric. }
\item{newrel_m65}{The variable is numeric. }
\item{newrel_m15plus}{The variable is numeric. }
\item{newrel_mu}{The variable is numeric. }
\item{newrel_f04}{The variable is numeric. }
\item{newrel_f514}{The variable is numeric. }
\item{newrel_f014}{The variable is numeric. }
\item{newrel_f1524}{The variable is numeric. }
\item{newrel_f2534}{The variable is numeric. }
\item{newrel_f3544}{The variable is numeric. }
\item{newrel_f4554}{The variable is numeric. }
\item{newrel_f5564}{The variable is numeric. }
\item{newrel_f65}{The variable is numeric. }
\item{newrel_f15plus}{The variable is numeric. }
\item{newrel_fu}{The variable is numeric. }
\item{newrel_sexunk04}{The variable is numeric. }
\item{newrel_sexunk514}{The variable is numeric. }
\item{newrel_sexunk014}{The variable is numeric. }
\item{newrel_sexunk15plus}{The variable is numeric. }
\item{newrel_sexunkageunk}{The variable is numeric. }
\item{rdx_data_available}{The variable is numeric. }
\item{newinc_rdx}{The variable is numeric. }
\item{rdxsurvey_newinc}{The variable is logical. }
\item{rdxsurvey_newinc_rdx}{The variable is logical. }
\item{rdst_new}{The variable is numeric. }
\item{rdst_ret}{The variable is numeric. }
\item{rdst_unk}{The variable is numeric. }
\item{conf_rrmdr}{The variable is numeric. }
\item{conf_mdr}{The variable is numeric. }
\item{rr_sldst}{The variable is numeric. }
\item{all_conf_xdr}{The variable is numeric. }
\item{unconf_rrmdr_tx}{The variable is numeric. }
\item{conf_rrmdr_tx}{The variable is numeric. }
\item{unconf_mdr_tx}{The variable is numeric. }
\item{conf_mdr_tx}{The variable is numeric. }
\item{conf_xdr_tx}{The variable is numeric. }
\item{mdrxdr_bdq_used}{The variable is numeric. }
\item{mdrxdr_bdq_tx}{The variable is numeric. }
\item{mdrxdr_dlm_used}{The variable is numeric. }
\item{mdrxdr_dlm_tx}{The variable is numeric. }
\item{mdr_shortreg_used}{The variable is numeric. }
\item{mdr_shortreg_tx}{The variable is numeric. }
\item{mdr_tx_adverse_events}{The variable is numeric. }
\item{mdr_tx_adsm}{The variable is numeric. }
\item{newrel_tbhiv_flg}{The variable is numeric. }
\item{newrel_hivtest}{The variable is numeric. }
\item{newrel_hivpos}{The variable is numeric. }
\item{newrel_art}{The variable is numeric. }
\item{hivtest}{The variable is numeric. }
\item{hivtest_pos}{The variable is numeric. }
\item{hiv_cpt}{The variable is numeric. }
\item{hiv_art}{The variable is numeric. }
\item{hiv_tbscr}{The variable is numeric. }
\item{hiv_reg}{The variable is numeric. }
\item{hiv_ipt}{The variable is numeric. }
\item{hiv_reg_new}{The variable is numeric. }
\item{hiv_ipt_reg_all}{The variable is numeric. }
\item{hiv_reg_all}{The variable is numeric. }
\item{hiv_tbdetect}{The variable is numeric. }
\item{hiv_reg_new2}{The variable is numeric. }
}
}
\source{
\url{https://www.who.int/tb/country/data/download/en/}
}
\usage{
tb_cases
}
\description{
See source for description of the data. tb_dictionary describes the column names.
}
\examples{
\dontrun{
tb_cases
}
}
\keyword{datasets}
