% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb_dictionary}
\alias{tb_dictionary}
\title{Word Health Organization (WHO) Tuberculosis csv file column names}
\format{
A data frame with columns:
\describe{
\item{variable_name}{The variable is character. }
\item{dataset}{The variable is character. }
\item{code_list}{The variable is character. }
\item{definition}{The variable is character. }
}
}
\source{
\url{https://www.who.int/tb/country/data/download/en/}
}
\usage{
tb_dictionary
}
\description{
File found at https://extranet.who.int/tme/generateCSV.asp?ds=dictionary
}
\examples{
\dontrun{
tb_dictionary
}
}
\keyword{datasets}
