% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb_estimates}
\alias{tb_estimates}
\title{Word Health Organization (WHO) Tuberculosis csv file column names}
\format{
A data frame with columns:
\describe{
\item{country}{The variable is character. }
\item{iso2}{The variable is character. }
\item{iso3}{The variable is character. }
\item{iso_numeric}{The variable is character. }
\item{g_whoregion}{The variable is character. }
\item{year}{The variable is numeric. }
\item{e_pop_num}{The variable is numeric. }
\item{e_inc_100k}{The variable is numeric. }
\item{e_inc_100k_lo}{The variable is numeric. }
\item{e_inc_100k_hi}{The variable is numeric. }
\item{e_inc_num}{The variable is numeric. }
\item{e_inc_num_lo}{The variable is numeric. }
\item{e_inc_num_hi}{The variable is numeric. }
\item{e_tbhiv_prct}{The variable is numeric. }
\item{e_tbhiv_prct_lo}{The variable is numeric. }
\item{e_tbhiv_prct_hi}{The variable is numeric. }
\item{e_inc_tbhiv_100k}{The variable is numeric. }
\item{e_inc_tbhiv_100k_lo}{The variable is numeric. }
\item{e_inc_tbhiv_100k_hi}{The variable is numeric. }
\item{e_inc_tbhiv_num}{The variable is numeric. }
\item{e_inc_tbhiv_num_lo}{The variable is numeric. }
\item{e_inc_tbhiv_num_hi}{The variable is numeric. }
\item{e_mort_exc_tbhiv_100k}{The variable is numeric. }
\item{e_mort_exc_tbhiv_100k_lo}{The variable is numeric. }
\item{e_mort_exc_tbhiv_100k_hi}{The variable is numeric. }
\item{e_mort_exc_tbhiv_num}{The variable is numeric. }
\item{e_mort_exc_tbhiv_num_lo}{The variable is numeric. }
\item{e_mort_exc_tbhiv_num_hi}{The variable is numeric. }
\item{e_mort_tbhiv_100k}{The variable is numeric. }
\item{e_mort_tbhiv_100k_lo}{The variable is numeric. }
\item{e_mort_tbhiv_100k_hi}{The variable is numeric. }
\item{e_mort_tbhiv_num}{The variable is numeric. }
\item{e_mort_tbhiv_num_lo}{The variable is numeric. }
\item{e_mort_tbhiv_num_hi}{The variable is numeric. }
\item{e_mort_100k}{The variable is numeric. }
\item{e_mort_100k_lo}{The variable is numeric. }
\item{e_mort_100k_hi}{The variable is numeric. }
\item{e_mort_num}{The variable is numeric. }
\item{e_mort_num_lo}{The variable is numeric. }
\item{e_mort_num_hi}{The variable is numeric. }
\item{cfr}{The variable is numeric. }
\item{cfr_lo}{The variable is numeric. }
\item{cfr_hi}{The variable is numeric. }
\item{cfr_pct}{The variable is numeric. }
\item{cfr_pct_lo}{The variable is numeric. }
\item{cfr_pct_hi}{The variable is numeric. }
\item{c_newinc_100k}{The variable is numeric. }
\item{c_cdr}{The variable is numeric. }
\item{c_cdr_lo}{The variable is numeric. }
\item{c_cdr_hi}{The variable is numeric. }
}
}
\source{
\url{https://www.who.int/tb/country/data/download/en/}
}
\usage{
tb_estimates
}
\description{
File found at https://extranet.who.int/tme/generateCSV.asp?ds=dictionary
}
\examples{
\dontrun{
tb_estimates
}
}
\keyword{datasets}
