% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb_outcomes}
\alias{tb_outcomes}
\title{Word Health Organization (WHO) Tuberculosis treatment outcomes by country}
\format{
A data frame with columns:
\describe{
\item{country}{The variable is character. }
\item{iso2}{The variable is character. }
\item{iso3}{The variable is character. }
\item{iso_numeric}{The variable is character. }
\item{g_whoregion}{The variable is character. }
\item{year}{The variable is numeric. }
\item{rep_meth}{The variable is numeric. }
\item{new_sp_coh}{The variable is numeric. }
\item{new_sp_cur}{The variable is numeric. }
\item{new_sp_cmplt}{The variable is numeric. }
\item{new_sp_died}{The variable is numeric. }
\item{new_sp_fail}{The variable is numeric. }
\item{new_sp_def}{The variable is numeric. }
\item{c_new_sp_tsr}{The variable is numeric. }
\item{new_snep_coh}{The variable is numeric. }
\item{new_snep_cmplt}{The variable is numeric. }
\item{new_snep_died}{The variable is numeric. }
\item{new_snep_fail}{The variable is numeric. }
\item{new_snep_def}{The variable is numeric. }
\item{c_new_snep_tsr}{The variable is numeric. }
\item{ret_coh}{The variable is numeric. }
\item{ret_cur}{The variable is numeric. }
\item{ret_cmplt}{The variable is numeric. }
\item{ret_died}{The variable is numeric. }
\item{ret_fail}{The variable is numeric. }
\item{ret_def}{The variable is numeric. }
\item{hiv_new_sp_coh}{The variable is numeric. }
\item{hiv_new_sp_cur}{The variable is numeric. }
\item{hiv_new_sp_cmplt}{The variable is numeric. }
\item{hiv_new_sp_died}{The variable is numeric. }
\item{hiv_new_sp_fail}{The variable is numeric. }
\item{hiv_new_sp_def}{The variable is numeric. }
\item{hiv_new_snep_coh}{The variable is numeric. }
\item{hiv_new_snep_cmplt}{The variable is numeric. }
\item{hiv_new_snep_died}{The variable is numeric. }
\item{hiv_new_snep_fail}{The variable is numeric. }
\item{hiv_new_snep_def}{The variable is numeric. }
\item{hiv_ret_coh}{The variable is numeric. }
\item{hiv_ret_cur}{The variable is numeric. }
\item{hiv_ret_cmplt}{The variable is numeric. }
\item{hiv_ret_died}{The variable is numeric. }
\item{hiv_ret_fail}{The variable is numeric. }
\item{hiv_ret_def}{The variable is numeric. }
\item{rel_with_new_flg}{The variable is numeric. }
\item{newrel_coh}{The variable is numeric. }
\item{newrel_succ}{The variable is numeric. }
\item{newrel_fail}{The variable is numeric. }
\item{newrel_died}{The variable is numeric. }
\item{newrel_lost}{The variable is numeric. }
\item{c_new_tsr}{The variable is numeric. }
\item{ret_nrel_coh}{The variable is numeric. }
\item{ret_nrel_succ}{The variable is numeric. }
\item{ret_nrel_fail}{The variable is numeric. }
\item{ret_nrel_died}{The variable is numeric. }
\item{ret_nrel_lost}{The variable is numeric. }
\item{c_ret_tsr}{The variable is numeric. }
\item{tbhiv_coh}{The variable is numeric. }
\item{tbhiv_succ}{The variable is numeric. }
\item{tbhiv_fail}{The variable is numeric. }
\item{tbhiv_died}{The variable is numeric. }
\item{tbhiv_lost}{The variable is numeric. }
\item{c_tbhiv_tsr}{The variable is numeric. }
\item{mdr_coh}{The variable is numeric. }
\item{mdr_succ}{The variable is numeric. }
\item{mdr_fail}{The variable is numeric. }
\item{mdr_died}{The variable is numeric. }
\item{mdr_lost}{The variable is numeric. }
\item{xdr_coh}{The variable is numeric. }
\item{xdr_succ}{The variable is numeric. }
\item{xdr_fail}{The variable is numeric. }
\item{xdr_died}{The variable is numeric. }
\item{xdr_lost}{The variable is numeric. }
}
}
\source{
\url{https://www.who.int/tb/country/data/download/en/}
}
\usage{
tb_outcomes
}
\description{
See source for description of the data. tb_dictionary describes the column names.
}
\examples{
\dontrun{
tb_outcomes
}
}
\keyword{datasets}
