% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb_rfds}
\alias{tb_rfds}
\title{Word Health Organization (WHO) Tuberculosis as found in the R for Data Science Book}
\format{
A data frame with columns:
\describe{
\item{country}{The variable is character. }
\item{iso2}{The variable is character. }
\item{iso3}{The variable is character. }
\item{year}{The variable is integer. }
\item{new_sp_m014}{The variable is integer. }
\item{new_sp_m1524}{The variable is integer. }
\item{new_sp_m2534}{The variable is integer. }
\item{new_sp_m3544}{The variable is integer. }
\item{new_sp_m4554}{The variable is integer. }
\item{new_sp_m5564}{The variable is integer. }
\item{new_sp_m65}{The variable is integer. }
\item{new_sp_f014}{The variable is integer. }
\item{new_sp_f1524}{The variable is integer. }
\item{new_sp_f2534}{The variable is integer. }
\item{new_sp_f3544}{The variable is integer. }
\item{new_sp_f4554}{The variable is integer. }
\item{new_sp_f5564}{The variable is integer. }
\item{new_sp_f65}{The variable is integer. }
\item{new_sn_m014}{The variable is integer. }
\item{new_sn_m1524}{The variable is integer. }
\item{new_sn_m2534}{The variable is integer. }
\item{new_sn_m3544}{The variable is integer. }
\item{new_sn_m4554}{The variable is integer. }
\item{new_sn_m5564}{The variable is integer. }
\item{new_sn_m65}{The variable is integer. }
\item{new_sn_f014}{The variable is integer. }
\item{new_sn_f1524}{The variable is integer. }
\item{new_sn_f2534}{The variable is integer. }
\item{new_sn_f3544}{The variable is integer. }
\item{new_sn_f4554}{The variable is integer. }
\item{new_sn_f5564}{The variable is integer. }
\item{new_sn_f65}{The variable is integer. }
\item{new_ep_m014}{The variable is integer. }
\item{new_ep_m1524}{The variable is integer. }
\item{new_ep_m2534}{The variable is integer. }
\item{new_ep_m3544}{The variable is integer. }
\item{new_ep_m4554}{The variable is integer. }
\item{new_ep_m5564}{The variable is integer. }
\item{new_ep_m65}{The variable is integer. }
\item{new_ep_f014}{The variable is integer. }
\item{new_ep_f1524}{The variable is integer. }
\item{new_ep_f2534}{The variable is integer. }
\item{new_ep_f3544}{The variable is integer. }
\item{new_ep_f4554}{The variable is integer. }
\item{new_ep_f5564}{The variable is integer. }
\item{new_ep_f65}{The variable is integer. }
\item{newrel_m014}{The variable is integer. }
\item{newrel_m1524}{The variable is integer. }
\item{newrel_m2534}{The variable is integer. }
\item{newrel_m3544}{The variable is integer. }
\item{newrel_m4554}{The variable is integer. }
\item{newrel_m5564}{The variable is integer. }
\item{newrel_m65}{The variable is integer. }
\item{newrel_f014}{The variable is integer. }
\item{newrel_f1524}{The variable is integer. }
\item{newrel_f2534}{The variable is integer. }
\item{newrel_f3544}{The variable is integer. }
\item{newrel_f4554}{The variable is integer. }
\item{newrel_f5564}{The variable is integer. }
\item{newrel_f65}{The variable is integer. }
}
}
\source{
\url{https://tidyr.tidyverse.org/reference/who.html}
}
\usage{
tb_rfds
}
\description{
See the tidy data chapter https://r4ds.had.co.nz/tidy-data.html
}
\examples{
\dontrun{
tb_rfds
}
}
\keyword{datasets}
