% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb_utilization}
\alias{tb_utilization}
\title{Word Health Organization (WHO) Tuberculosis expenditures and utilization by country}
\format{
A data frame with columns:
\describe{
\item{country}{The variable is character. }
\item{iso2}{The variable is character. }
\item{iso3}{The variable is character. }
\item{iso_numeric}{The variable is character. }
\item{g_whoregion}{The variable is character. }
\item{year}{The variable is numeric. }
\item{exp_cpp_dstb}{The variable is numeric. }
\item{exp_cpp_mdr}{The variable is numeric. }
\item{exp_cpp_xdr}{The variable is numeric. }
\item{exp_lab}{The variable is numeric. }
\item{rcvd_lab}{The variable is numeric. }
\item{exp_staff}{The variable is numeric. }
\item{rcvd_staff}{The variable is numeric. }
\item{exp_fld}{The variable is numeric. }
\item{rcvd_fld}{The variable is numeric. }
\item{exp_prog}{The variable is numeric. }
\item{rcvd_prog}{The variable is numeric. }
\item{exp_sld}{The variable is numeric. }
\item{rcvd_sld}{The variable is numeric. }
\item{exp_mdrmgt}{The variable is numeric. }
\item{rcvd_mdrmgt}{The variable is numeric. }
\item{exp_tbhiv}{The variable is numeric. }
\item{rcvd_tbhiv}{The variable is numeric. }
\item{exp_patsup}{The variable is numeric. }
\item{rcvd_patsup}{The variable is numeric. }
\item{exp_orsrvy}{The variable is numeric. }
\item{rcvd_orsrvy}{The variable is numeric. }
\item{exp_oth}{The variable is numeric. }
\item{rcvd_oth}{The variable is numeric. }
\item{exp_tot}{The variable is numeric. }
\item{rcvd_tot}{The variable is numeric. }
\item{rcvd_tot_domestic}{The variable is numeric. }
\item{rcvd_tot_gf}{The variable is numeric. }
\item{rcvd_tot_usaid}{The variable is numeric. }
\item{rcvd_tot_grnt}{The variable is numeric. }
\item{rcvd_tot_sources}{The variable is numeric. }
\item{hcfvisit_dstb}{The variable is numeric. }
\item{hcfvisit_mdr}{The variable is numeric. }
\item{hospd_dstb_prct}{The variable is numeric. }
\item{hospd_mdr_prct}{The variable is numeric. }
\item{hospd_dstb_dur}{The variable is numeric. }
\item{hospd_mdr_dur}{The variable is numeric. }
\item{hosp_type_mdr}{The variable is numeric. }
}
}
\source{
\url{https://www.who.int/tb/country/data/download/en/}
}
\usage{
tb_utilization
}
\description{
See source for description of the data. tb_dictionary describes the column names.
}
\examples{
\dontrun{
tb_utilization
}
}
\keyword{datasets}
